+++ statutory/lib/metrics.js
@@ -19,27 +19,44 @@ const createGauge = (name, help, labels = []) => new Gauge({
 });
 
 const gaugesList = {
-    eventsTotal: createGauge('events_total', 'Total amount of statutory events'),
-    eventsByType: createGauge('events_by_type', 'Amount of events by type', ['type']),
-    eventsByTypeAndStatus: createGauge('events_by_type_and_status', 'Amount of events by type and status', ['type', 'status']),
-    applicationsTotal: createGauge('applications_total', 'Total amount of applications'),
-    applicationsByEvent: createGauge('applications_by_event', 'Amount of applications by event', ['event_name']),
-    applicationsByEventAndBody: createGauge('applications_by_event_and_body', 'Amount of applications by event and body', ['event_name', 'body_name']),
+    eventsTotal: createGauge('statutory_events_total', 'Total amount of statutory events'),
+    eventsByType: createGauge('statutory_events_by_type', 'Amount of statutory events by type', ['type']),
+    eventsByTypeAndStatus: createGauge(
+        'statutory_events_by_type_and_status',
+        'Amount of statutory events by type and status',
+        ['type', 'status']
+    ),
+    applicationsTotal: createGauge('statutory_applications_total', 'Total amount of statutory applications'),
+    applicationsByEvent: createGauge('statutory_applications_by_event', 'Amount of statutory applications by event', ['event_name']),
+    applicationsByEventAndBody: createGauge(
+        'statutory_applications_by_event_and_body',
+        'Amount of statutory applications by event and body',
+        ['event_name', 'body_name']
+    ),
     applicationsByEventAndPaxType: createGauge(
         'applications_by_event_and_pax_type',
-        'Amount of applications by event and pax type',
+        'Amount of statutory applications by event and pax type',
         ['event_name', 'participant_type']
     ),
-    memberslistsTotal: createGauge('memberslists_total', 'Total amount of memberslists'),
-    memberslistsByEvent: createGauge('memberslists_by_event', 'Amount of memberslists by event', ['event_name']),
-    memberslistsByEventAndBody: createGauge('memberslists_by_event_and_body', 'Amount of memberslists by event and body', ['event_name', 'body_id']),
-    positionsTotal: createGauge('positions_total', 'Total amount of positions'),
-    positionsByEvent: createGauge('positions_by_event', 'Amount of positions by event', ['event_name']),
-    candidatesTotal: createGauge('candidates_total', 'Total amount of candidates'),
-    candidatesByStatus: createGauge('candidates_by_status', 'Amount of candidates by status', ['status']),
+    memberslistsTotal: createGauge('statutory_memberslists_total', 'Total amount of statutory memberslists'),
+    memberslistsByEvent: createGauge('statutory_memberslists_by_event', 'Amount of statutory memberslists by event', ['event_name']),
+    memberslistsByEventAndBody: createGauge(
+        'statutory_memberslists_by_event_and_body',
+        'Amount of statutory memberslists by event and body',
+        ['event_name', 'body_id']
+    ),
+    positionsTotal: createGauge('statutory_positions_total', 'Total amount of statutory positions'),
+    positionsByEvent: createGauge('statutory_positions_by_event', 'Amount of statutory positions by event', ['event_name']),
+    candidatesTotal: createGauge('statutory_candidates_total', 'Total amount of statutory candidates'),
+    candidatesByStatus: createGauge('statutory_candidates_by_status', 'Amount of statutory candidates by status', ['status']),
+    candidatesByEventAndPosition: createGauge(
+        'statutory_candidates_by_event_and_position',
+        'Amount of statutory candidates by event and position',
+        ['event_name', 'position_name']
+    ),
     candidatesByStatusAndPosition: createGauge(
-        'candidates_by_status_and_position',
-        'Amount of candidates by status and position',
+        'statutory_candidates_by_status_and_position',
+        'Amount of statutory candidates by status and position',
         ['status', 'position_name', 'event_name']
     ),
 
@@ -98,6 +115,7 @@ exports.getMetrics = async (req, res) => {
 
     helpers.addGaugeData(gaugesList.candidatesTotal, helpers.countByFields(candidates, []));
     helpers.addGaugeData(gaugesList.candidatesByStatus, helpers.countByFields(candidates, ['status']));
+    helpers.addGaugeData(gaugesList.candidatesByEventAndPosition, helpers.countByFields(candidates, ['position_name', 'event_name']));
     helpers.addGaugeData(gaugesList.candidatesByStatusAndPosition, helpers.countByFields(candidates, ['status', 'position_name', 'event_name']));
 
     res.set('Content-Type', register.contentType);